data_management/preprocessing/quality_control.py
#!/usr/bin/env python3
"""
数据质量控制模块

实现多源数据的质量控制，包括：
- 异常值检测和处理
- 数据完整性检查
- 多源数据交叉验证
- 质量评级和标记

Author: Tibetan Glacier PINNs Project Team
Date: 2025
"""

import numpy as np
import xarray as xr
import pandas as pd
from pathlib import Path
import logging
from typing import Dict, List, Optional, Tuple, Union
from scipy import stats
from sklearn.ensemble import IsolationForest
from sklearn.preprocessing import StandardScaler

# 配置日志
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class QualityController:
    """
    数据质量控制器
    
    对多源冰川数据进行质量控制和验证
    """
    
    def __init__(self, quality_thresholds: Optional[Dict[str, float]] = None):
        """
        初始化质量控制器
        
        Args:
            quality_thresholds: 质量阈值字典
        """
        self.quality_thresholds = quality_thresholds or {
            'outlier_zscore': 3.0,
            'missing_data_ratio': 0.3,
            'correlation_threshold': 0.7,
            'physical_bounds_violation': 0.05
        }
        
    def detect_outliers_statistical(self, data: xr.DataArray, method: str = 'zscore') -> xr.DataArray:
        """
        统计方法检测异常值
        
        Args:
            data: 输入数据
            method: 检测方法 ('zscore', 'iqr', 'isolation_forest')
            
        Returns:
            DataArray: 异常值掩码 (True=异常值)
        """
        # TODO: 实现统计异常值检测
        pass
        
    def detect_outliers_physical(self, data: xr.DataArray, data_type: str) -> xr.DataArray:
        """
        物理约束检测异常值
        
        Args:
            data: 输入数据
            data_type: 数据类型 ('thickness', 'velocity', 'mass_balance')
            
        Returns:
            DataArray: 异常值掩码
        """
        # TODO: 实现物理约束异常值检测
        pass
        
    def check_data_completeness(self, data: xr.DataArray) -> Dict[str, float]:
        """
        检查数据完整性
        
        Args:
            data: 输入数据
            
        Returns:
            Dict: 完整性统计
        """
        # TODO: 实现数据完整性检查
        pass
        
    def validate_temporal_consistency(self, time_series: xr.DataArray) -> Dict[str, Union[bool, float]]:
        """
        验证时间一致性
        
        Args:
            time_series: 时间序列数据
            
        Returns:
            Dict: 时间一致性验证结果
        """
        # TODO: 实现时间一致性验证
        pass
        
    def cross_validate_sources(self, data_dict: Dict[str, xr.DataArray]) -> Dict[str, Dict[str, float]]:
        """
        多源数据交叉验证
        
        Args:
            data_dict: 多源数据字典
            
        Returns:
            Dict: 交叉验证结果
        """
        # TODO: 实现多源数据交叉验证
        pass
        
    def validate_physical_constraints(self, data: xr.DataArray, data_type: str) -> Dict[str, bool]:
        """
        验证物理约束
        
        Args:
            data: 输入数据
            data_type: 数据类型
            
        Returns:
            Dict: 物理约束验证结果
        """
        # TODO: 实现物理约束验证
        pass
        
    def calculate_uncertainty_metrics(self, data: xr.DataArray) -> Dict[str, float]:
        """
        计算不确定性指标
        
        Args:
            data: 输入数据
            
        Returns:
            Dict: 不确定性指标
        """
        # TODO: 实现不确定性指标计算
        pass
        
    def assign_quality_grades(self, data: xr.DataArray, 
                            validation_results: Dict[str, Union[bool, float]]) -> xr.DataArray:
        """
        分配质量等级
        
        Args:
            data: 输入数据
            validation_results: 验证结果
            
        Returns:
            DataArray: 质量等级数组
        """
        # TODO: 实现质量等级分配
        pass
        
    def filter_by_quality(self, data: xr.DataArray, quality_grades: xr.DataArray, 
                         min_grade: str = 'B') -> xr.DataArray:
        """
        按质量等级过滤数据
        
        Args:
            data: 输入数据
            quality_grades: 质量等级
            min_grade: 最低质量等级
            
        Returns:
            DataArray: 过滤后的数据
        """
        # TODO: 实现质量过滤
        pass
        
    def generate_quality_flags(self, data: xr.DataArray, data_type: str) -> xr.Dataset:
        """
        生成质量标记
        
        Args:
            data: 输入数据
            data_type: 数据类型
            
        Returns:
            Dataset: 包含各种质量标记的数据集
        """
        # TODO: 实现质量标记生成
        pass
        
    def comprehensive_quality_control(self, data_dict: Dict[str, Dict[str, xr.DataArray]]) -> Dict[str, Dict[str, xr.DataArray]]:
        """
        综合质量控制
        
        Args:
            data_dict: 多源数据字典
            
        Returns:
            Dict: 质量控制后的数据
        """
        logger.info("开始综合数据质量控制...")
        
        quality_controlled_data = {}
        quality_reports = {}
        
        for source_name, source_data in data_dict.items():
            logger.info(f"质量控制{source_name}数据...")
            
            qc_source = {}
            source_reports = {}
            
            for var_name, var_data in source_data.items():
                # 统计异常值检测
                statistical_outliers = self.detect_outliers_statistical(var_data)
                
                # 物理约束异常值检测
                physical_outliers = self.detect_outliers_physical(var_data, var_name)
                
                # 数据完整性检查
                completeness = self.check_data_completeness(var_data)
                
                # 时间一致性验证
                if 'time' in var_data.dims:
                    temporal_consistency = self.validate_temporal_consistency(var_data)
                else:
                    temporal_consistency = {'consistent': True}
                
                # 物理约束验证
                physical_constraints = self.validate_physical_constraints(var_data, var_name)
                
                # 不确定性指标
                uncertainty_metrics = self.calculate_uncertainty_metrics(var_data)
                
                # 综合验证结果
                validation_results = {
                    **completeness,
                    **temporal_consistency,
                    **physical_constraints,
                    **uncertainty_metrics
                }
                
                # 分配质量等级
                quality_grades = self.assign_quality_grades(var_data, validation_results)
                
                # 生成质量标记
                quality_flags = self.generate_quality_flags(var_data, var_name)
                
                # 按质量过滤数据
                filtered_data = self.filter_by_quality(var_data, quality_grades)
                
                qc_source[var_name] = filtered_data
                qc_source[f'{var_name}_quality_grades'] = quality_grades
                qc_source[f'{var_name}_quality_flags'] = quality_flags
                
                source_reports[var_name] = validation_results
            
            quality_controlled_data[source_name] = qc_source
            quality_reports[source_name] = source_reports
        
        # 多源交叉验证
        cross_validation_results = self.cross_validate_sources(
            {k: {vk: vv for vk, vv in v.items() if not vk.endswith(('_quality_grades', '_quality_flags'))}
             for k, v in quality_controlled_data.items()}
        )
        
        logger.info("综合数据质量控制完成")
        logger.info(f"交叉验证结果: {cross_validation_results}")
        
        return quality_controlled_data
        
    def generate_quality_report(self, quality_controlled_data: Dict[str, Dict[str, xr.DataArray]]) -> Dict[str, Dict[str, Union[str, float, bool]]]:
        """
        生成质量控制报告
        
        Args:
            quality_controlled_data: 质量控制后的数据
            
        Returns:
            Dict: 质量控制报告
        """
        # TODO: 实现质量控制报告生成
        pass

if __name__ == "__main__":
    # 示例用法
    qc = QualityController()
    # quality_controlled_data = qc.comprehensive_quality_control(data_dict)