main_experiment.py
#!/usr/bin/env python
# -*- coding: utf-8 -*-
"""
青藏高原冰川PINNs建模主实验脚本
Tibetan Plateau Glacier PINNs Modeling - Main Experiment Script

本脚本是整个项目的主入口，负责协调数据处理、模型训练、验证和分析的完整流程。
"""

import os
import sys
import argparse
import logging
from pathlib import Path
from typing import Dict, Any, Optional
import yaml
import torch
import numpy as np
from datetime import datetime

# 添加项目根目录到Python路径
project_root = Path(__file__).parent
sys.path.insert(0, str(project_root))

# 导入项目模块（注意：这些模块需要在后续步骤中实现）
# from data_management.preprocessing import DataPreprocessor
# from model_architecture.core_pinns import BasePINN
# from training_framework.training_stages import ProgressiveTrainer
# from validation_testing.physics_validation import PhysicsValidator
# from analysis_visualization.spatial_analysis import SpatialAnalyzer

def setup_logging(log_level: str = "INFO", log_file: Optional[str] = None) -> None:
    """
    设置日志配置
    
    Args:
        log_level: 日志级别
        log_file: 日志文件路径
    """
    log_format = "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
    
    if log_file:
        logging.basicConfig(
            level=getattr(logging, log_level.upper()),
            format=log_format,
            handlers=[
                logging.FileHandler(log_file),
                logging.StreamHandler(sys.stdout)
            ]
        )
    else:
        logging.basicConfig(
            level=getattr(logging, log_level.upper()),
            format=log_format,
            stream=sys.stdout
        )

def load_config(config_path: str) -> Dict[str, Any]:
    """
    加载实验配置文件
    
    Args:
        config_path: 配置文件路径
        
    Returns:
        Dict: 配置参数字典
    """
    try:
        with open(config_path, 'r', encoding='utf-8') as f:
            config = yaml.safe_load(f)
        return config
    except FileNotFoundError:
        logging.error(f"配置文件未找到: {config_path}")
        return {}
    except yaml.YAMLError as e:
        logging.error(f"配置文件解析错误: {e}")
        return {}

def setup_environment(config: Dict[str, Any]) -> None:
    """
    设置实验环境
    
    Args:
        config: 配置参数
    """
    # 设置随机种子
    seed = config.get('random_seed', 42)
    np.random.seed(seed)
    torch.manual_seed(seed)
    if torch.cuda.is_available():
        torch.cuda.manual_seed(seed)
        torch.cuda.manual_seed_all(seed)
    
    # 设置CUDA设备
    if torch.cuda.is_available():
        device_ids = config.get('gpu_ids', [0])
        os.environ['CUDA_VISIBLE_DEVICES'] = ','.join(map(str, device_ids))
        logging.info(f"使用GPU设备: {device_ids}")
    else:
        logging.warning("CUDA不可用，将使用CPU进行计算")
    
    # 创建实验目录
    experiment_name = config.get('experiment_name', f"experiment_{datetime.now().strftime('%Y%m%d_%H%M%S')}")
    experiment_dir = project_root / "experiments" / "results" / experiment_name
    experiment_dir.mkdir(parents=True, exist_ok=True)
    
    config['experiment_dir'] = str(experiment_dir)
    logging.info(f"实验目录: {experiment_dir}")

def run_data_preprocessing(config: Dict[str, Any]) -> bool:
    """
    运行数据预处理阶段
    
    Args:
        config: 配置参数
        
    Returns:
        bool: 预处理是否成功
    """
    logging.info("开始数据预处理阶段...")
    
    try:
        # TODO: 实现数据预处理逻辑
        # preprocessor = DataPreprocessor(config)
        # preprocessor.process_all_datasets()
        
        logging.info("数据预处理完成")
        return True
    except Exception as e:
        logging.error(f"数据预处理失败: {e}")
        return False

def run_model_training(config: Dict[str, Any]) -> bool:
    """
    运行模型训练阶段
    
    Args:
        config: 配置参数
        
    Returns:
        bool: 训练是否成功
    """
    logging.info("开始模型训练阶段...")
    
    try:
        # TODO: 实现模型训练逻辑
        # trainer = ProgressiveTrainer(config)
        # trainer.run_three_stage_training()
        
        logging.info("模型训练完成")
        return True
    except Exception as e:
        logging.error(f"模型训练失败: {e}")
        return False

def run_model_validation(config: Dict[str, Any]) -> bool:
    """
    运行模型验证阶段
    
    Args:
        config: 配置参数
        
    Returns:
        bool: 验证是否成功
    """
    logging.info("开始模型验证阶段...")
    
    try:
        # TODO: 实现模型验证逻辑
        # validator = PhysicsValidator(config)
        # validator.run_comprehensive_validation()
        
        logging.info("模型验证完成")
        return True
    except Exception as e:
        logging.error(f"模型验证失败: {e}")
        return False

def run_result_analysis(config: Dict[str, Any]) -> bool:
    """
    运行结果分析阶段
    
    Args:
        config: 配置参数
        
    Returns:
        bool: 分析是否成功
    """
    logging.info("开始结果分析阶段...")
    
    try:
        # TODO: 实现结果分析逻辑
        # analyzer = SpatialAnalyzer(config)
        # analyzer.generate_comprehensive_analysis()
        
        logging.info("结果分析完成")
        return True
    except Exception as e:
        logging.error(f"结果分析失败: {e}")
        return False

def main():
    """
    主函数
    """
    parser = argparse.ArgumentParser(description="青藏高原冰川PINNs建模主实验")
    parser.add_argument(
        "--config", 
        type=str, 
        default="experiments/experiment_configs/default_config.yml",
        help="实验配置文件路径"
    )
    parser.add_argument(
        "--stage", 
        type=str, 
        choices=["all", "preprocess", "train", "validate", "analyze"],
        default="all",
        help="运行阶段"
    )
    parser.add_argument(
        "--log-level", 
        type=str, 
        choices=["DEBUG", "INFO", "WARNING", "ERROR"],
        default="INFO",
        help="日志级别"
    )
    parser.add_argument(
        "--log-file", 
        type=str,
        help="日志文件路径"
    )
    
    args = parser.parse_args()
    
    # 设置日志
    setup_logging(args.log_level, args.log_file)
    
    # 加载配置
    config = load_config(args.config)
    if not config:
        logging.error("无法加载配置文件，退出程序")
        sys.exit(1)
    
    # 设置环境
    setup_environment(config)
    
    logging.info("="*50)
    logging.info("青藏高原冰川PINNs建模项目启动")
    logging.info(f"实验阶段: {args.stage}")
    logging.info(f"配置文件: {args.config}")
    logging.info("="*50)
    
    # 运行指定阶段
    success = True
    
    if args.stage in ["all", "preprocess"]:
        success &= run_data_preprocessing(config)
        if not success and args.stage == "preprocess":
            sys.exit(1)
    
    if args.stage in ["all", "train"] and success:
        success &= run_model_training(config)
        if not success and args.stage == "train":
            sys.exit(1)
    
    if args.stage in ["all", "validate"] and success:
        success &= run_model_validation(config)
        if not success and args.stage == "validate":
            sys.exit(1)
    
    if args.stage in ["all", "analyze"] and success:
        success &= run_result_analysis(config)
    
    if success:
        logging.info("实验成功完成！")
    else:
        logging.error("实验执行失败！")
        sys.exit(1)

if __name__ == "__main__":
    main()