requirements.txt
# Core Scientific Computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
xarray>=0.19.0
dask>=2021.6.0
numba>=0.54.0

# Machine Learning & Deep Learning
torch>=1.12.0
torchvision>=0.13.0
jax>=0.3.0
jaxlib>=0.3.0
optax>=0.1.0
pytorch-lightning>=1.6.0
scikit-learn>=1.0.0
statsmodels>=0.12.0
pymc3>=3.11.0

# Geospatial Data Processing
rasterio>=1.2.0
geopandas>=0.9.0
cartopy>=0.20.0
fiona>=1.8.0
shapely>=1.7.0
pyproj>=3.1.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
bokeh>=2.3.0
holoviews>=1.14.0
folium>=0.12.0
geoviews>=1.9.0

# Web Applications
streamlit>=1.0.0
dash>=2.0.0
fastapi>=0.68.0
uvicorn>=0.15.0

# Development & Testing
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.6.0
flake8>=3.9.0
mypy>=0.910
jupyter>=1.0.0
ipykernel>=6.0.0

# Configuration & Utilities
pyyaml>=5.4.0
toml>=0.10.0
tqdm>=4.61.0
click>=8.0.0
loguru>=0.5.0

# Optional: GPU Support
# torch-geometric>=2.0.0  # For graph neural networks
# cupy>=9.0.0  # For GPU acceleration
# rapids-cudf>=21.06  # For GPU dataframes